+++ app/src/Navbar/Navbar.js
@@ -2,6 +2,13 @@ import React from 'react';
 import './Navbar.scss';
 import logo from '../Resources/logo.png';
 import { Link, useLocation } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
+import { Route } from 'react-router-dom';
+import { Routes } from 'react-router-dom';
+import About from '../About/About';
+import Help from '../Help/Help';
+import App from '../App/App';
+import Feedback from '../Feedback/Feedback';
 
 /**
  * Navbar
@@ -11,10 +18,24 @@ import { Link, useLocation } from 'react-router-dom';
  */
 const Navbar = () => {
 
+  const handleHelpClick = () => {
+    
+  }
+
+  const handleAboutClick = () => {
+
+  }
+
+  const handleFeedbackClick = () => {
+
+  }
+
   const location = useLocation();
   return (
+    
     <nav className="navbar">
-      <div className="logo">
+      
+    <div className="logo">
         <div className='logo-image'>
           <Link to="/">
             <img 
@@ -31,13 +52,15 @@ const Navbar = () => {
       </div>
       
       <ul className="nav-links">
-        {location.pathname !== '/' && <li><Link to="/" className='link'> Home </Link></li>}
-        {location.pathname !== '/help' && <li> <Link to='/help' className='link'> Help </Link></li>}
-        {location.pathname !== '/about' && <li><Link to="/about" className='link'> About </Link></li>}
-        {location.pathname !== '/feedback' && <li> <Link to='/feedback' className='link'> Feedback </Link></li>}
+        {location.pathname !== '/' ? (<li><Link to="/" className='link'> Home </Link></li>) : (<></>)}
+        {location.pathname !== '/help' ?  (<li> <Link to='/help' className='link'> Help </Link></li>) : (<></>)}
+        {location.pathname !== '/about' ? (<li><Link to="/about" className='link'> About </Link></li>) : (<></>)}
+        {location.pathname !== '/feedback' ? (<li> <Link to='/feedback' className='link'> Feedback </Link></li>) : (<></>)}
       </ul>
     </nav>
   );
 };
+
+
 
 export default Navbar;
