+++ app/src/__tests__/Navbar.test.js
@@ -1,0 +1,11 @@
+import React from "react";
+import {fireEvent, render, screen} from '@testing-library/react';
+import CriticalAlerts from "../CriticalAlerts/CriticalAlerts";
+import '@testing-library/jest-dom/extend-expect';
+// import Navbar from '../Navbar/Navbar.js';
+import 'jest-canvas-mock';
+
+test('Simple render test', () => {
+    // window.URL.createObjectURL = function() {};
+    // render(<RoutedNavbar/>)
+})
