+++ app/src/index.js
@@ -13,7 +13,7 @@ root.render(
   <React.StrictMode>
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={ <App /> }> </Route>
+        <Route path="/" element={ <App/> }> </Route>
         <Route path="/about" element={<About/>}> </Route>
         <Route path='/help' element={<Help/>}> </Route>
         <Route path='/feedback' element={<Feedback/>}> </Route>
@@ -21,8 +21,3 @@ root.render(
     </BrowserRouter>
   </React.StrictMode>
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
